Snake 2.0 BombSquad
I have modified the original snake mod of bombsquad to give more mines. From this file you will get max 1000 of mines on the snake minigame. Normally in snake mod we get 120 mines. I faced a problem in team battle where the round ended so fast. So i had to increase the numbers of falling mines.i thought maybe there are more bs players who would need this file too.. The choices of numbers of mines in this mod are 'choices':[('Few', 160), ('Some', 280), ('Some more', 420), ('Many much', 740), ('wow', 1000). We get maximum of 1000 mines. If you need it to be modified more.. Or if you want me to mod some scripts drop a text.